FDG-8327 Move and adjust tests
